requirements.txt
boto==2.40.0
brotlipy==0.3.0
cffi==1.6.0
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-filter==0.13.0
django-storages==1.4.1
djangorestframework==3.3.3
Markdown==2.6.6
oauthlib==1.1.2
psycopg2==2.6.1
pycparser==2.14
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
whitenoise==3.2
